Handle tesselation error and switch to u32 index format

+++ src/render/options.rs
@@ -2,7 +2,7 @@ use wgpu::BufferAddress;
 
 pub const DEBUG_WIREFRAME: bool = false;
 pub const DEBUG_STENCIL_PATTERN: bool = false;
-pub const INDEX_FORMAT: wgpu::IndexFormat = wgpu::IndexFormat::Uint16; // Must match IndexDataType
+pub const INDEX_FORMAT: wgpu::IndexFormat = wgpu::IndexFormat::Uint32; // Must match IndexDataType
 
 pub const VERTEX_BUFFER_SIZE: BufferAddress = 1024 * 1024 * 32;
 pub const FEATURE_METADATA_BUFFER_SIZE: BufferAddress = 1024 * 1024 * 32;
